feat!(ari): go over struct derives and place them in separate files
- Place `Debug` first
- Request params derives `Serialize` and `Debug` only. Fields are
  all correctly serialized as "camelCase". (Rename all is sometimes
  superflously applied to avoid future compatibility issues.)
- Response data derives `Deserialize` only. Their fields are made
  private, exposed through `Getters` derive.
