+++ Cargo.toml
@@ -1,3 +1,9 @@
 [workspace]
 members = ["agirs", "amirs", "arirs"]
 resolver = "2"
+
+[workspace.lints.rust]
+unused_must_use = "deny"
+
+[workspace.lints.clippy]
+self_named_module_files = "deny"
